refactor: initialization script to index.ts, TS port. Fixes Spring embedded algorithm rendering bug

+++ src/lib/algorithms/springEmbedders/SpringEmbeddersGPUAlgorithm.js
@@ -226,6 +226,9 @@ export class SpringEmbeddersGPUAlgorithm {
   }
 
   onRemove() {
+    if (this._kernel === null) {
+      return; // already removed
+    }
     this._kernel.delete();
     this._kernel = null;
     if (this._nodesTexture === null) {
